Added Climber and Gripper Class. Made changes to Conveyer, Robot, and Shooter classes. 2-1-2020 -Jordan Migneault

+++ src/main/java/frc/robot/Shooter.java
@@ -14,7 +14,7 @@ public class Shooter
         this.motor2 = backwardMotor;
     }
 
-    public void Shoot(int power)
+    public void Shoot(double power)
     {
         motor1.set(ControlMode.PercentOutput, power);
         motor2.set(ControlMode.PercentOutput, -power);
